Setup demo Flask server for fetching latest pwv value
