+++ agents/pwv_agent/pwv_agent.py
@@ -1,7 +1,5 @@
 import numpy as np
 from numpy import random
-import datetime
-import time
 import os
 from os import environ
 
@@ -12,41 +10,12 @@ from autobahn.twisted.util import sleep as dsleep
 import argparse
 import txaio
 
+# temporary until Agent is updated to read from API
+from api.pwv_web import read_data_from_textfile
+
 # For logging
 txaio.use_twisted()
 LOG = txaio.make_logger()
-
-
-def julian_day_year_to_unixtime(day, year):
-    """
-    Convert water vapor radiometer's output Julian Day to unix timestamp.
-
-    Args:
-        day (float): day of the year
-        year (int):  year for the corresponding Julian Day
-    """
-    a = datetime.datetime(year, 1, 1) + datetime.timedelta(day-1)
-    unixtime = time.mktime(a.timetuple())
-
-    return unixtime
-
-
-def read_data_from_textfile(filename, year):
-    with open(filename, 'r') as f:
-        i = 0
-        for l in f.readlines():
-            if i == 0:
-                pass  # skip header
-            else:
-                line = l.strip().split()
-                timestamp = julian_day_year_to_unixtime(float(line[0]), year)
-
-                pwv = float(line[1])
-
-                _data = (pwv, timestamp)
-
-            i += 1
-        return _data
 
 
 class PWV_Agent:
